+++ deployments/chef/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Initial support for Splunk OpenTelemetry Auto Instrumentation for Java (Linux only)
+
 ## chef-v0.2.0
 
 - Initial release